Added testcases
