+++ activity_calendar/tests/commitee_pages/test_views.py
@@ -6,7 +6,7 @@ from django.views.generic import ListView, FormView
 
 from activity_calendar.models import Activity, OrganiserLink
 from committees.models import AssociationGroup
-from committees.views import AssociationGroupMixin
+from committees.committeecollective import AssociationGroupMixin
 from core.tests.util import suppress_warnings
 from utils.testing.view_test_utils import ViewValidityMixin
