+++ committees/tests/committee_pages/test_config.py
@@ -1,0 +1,18 @@
+from django.test import TestCase
+
+from committees.committeecollective import CommitteeConfig
+from committees.committee_pages.config import AssociationGroupWelcomeConfig
+
+from committees.committeecollective import registry
+
+
+class AssociationGroupHomeConfigTestCase(TestCase):
+
+    def test_class(self):
+        self.assertTrue(issubclass(AssociationGroupWelcomeConfig, CommitteeConfig))
+        self.assertEqual(AssociationGroupWelcomeConfig.url_keyword, "main")
+        self.assertEqual(AssociationGroupWelcomeConfig.url_name, "group_general")
+
+    def test_tab_order(self):
+        """ Tests that this config is first in the tab order """
+        self.assertIsInstance(registry.configs[0], AssociationGroupWelcomeConfig)
